requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.12.0
librosa
numpy
soundfile
joblib
scikit-learn
google-cloud-bigquery
google-cloud-storage
gcsfs
pandas
python-multipart